+++ packages/core/src/utilities/index.ts
@@ -17,6 +17,7 @@ import clamp from './clamp';
 import isOpposite from './isOpposite';
 import getClosestImageId from './getClosestImageId';
 import getSpacingInNormalDirection from './getSpacingInNormalDirection';
+import getEffectiveSpacingAlongDirection from './getEffectiveSpacingAlongDirection';
 import getInPlaneSpacingAndXYDirections from './getInPlaneSpacingAndXYDirections';
 import iterateOverPlane from './iterateOverPlane';
 import * as obliqueIntegerIterator from './obliqueIntegerIterator';
@@ -127,6 +128,7 @@ export type { ViewportVoiMappingProps } from './viewportVoiIntensityMapping';
 export * from './getPixelSpacingInformation';
 export * from './getPlaneCubeIntersectionDimensions';
 export * from './rotateToViewCoordinates';
+export * as voxelSlab from './voxelSlab';
 import { asArray } from './asArray';
 import {
   viewportSupportsImageSlices,
@@ -179,6 +181,7 @@ export {
   convertColorArrayToRgbString,
   getClosestImageId,
   getSpacingInNormalDirection,
+  getEffectiveSpacingAlongDirection,
   getInPlaneSpacingAndXYDirections,
   iterateOverPlane,
   obliqueIntegerIterator,