+++ packages/tools/src/stateManagement/segmentation/utilities/removeContourIslands.ts
@@ -4,6 +4,11 @@ import { getSegmentation } from '../getSegmentation';
 import { extractSegmentPolylines } from './extractSegmentPolylines';
 import findIslands from '../../../utilities/contours/findIslands';
 import { removeCompleteContourAnnotation } from './removeCompleteContourAnnotation';
+import { utilities as cornerstoneUtilities } from '@cornerstonejs/core';
+
+const cs3dLogger = cornerstoneUtilities.logger.toolsLog.getLogger(
+  'stateManagement.segmentation.utilities.removeContourIslands'
+);
 
 /**
  * Removes contour islands from a segmentation segment by detecting and deleting small isolated contours.
@@ -22,11 +27,11 @@ export default function removeContourIslands(
 ) {
   const segmentation = getSegmentation(segmentationId);
   if (!segmentation) {
-    console.warn(`Invalid segmentation given ${segmentationId}`);
+    cs3dLogger.warn(`Invalid segmentation given ${segmentationId}`);
     return;
   }
   if (!segmentation.representationData.Contour) {
-    console.warn(
+    cs3dLogger.warn(
       `No contour representation found for segmentation ${segmentationId}`
     );
     return;
@@ -37,7 +42,7 @@ export default function removeContourIslands(
     segmentIndex
   );
   if (!polylinesCanvasMap) {
-    console.warn(
+    cs3dLogger.warn(
       `Error extracting contour data from segment ${segmentIndex} in segmentation ${segmentationId}`
     );
     return;