+++ packages/core/CHANGELOG.md
@@ -3,6 +3,90 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.10.0](https://github.com/cornerstonejs/cornerstone3D/compare/v5.9.1...v5.10.0) (2026-09-10)
+
+### Features
+
+- **core:** define voxel-relative annotation statistics and add a shared oblique-capable voxel iterator ([#2893](https://github.com/cornerstonejs/cornerstone3D/issues/2893)) ([cd7f279](https://github.com/cornerstonejs/cornerstone3D/commit/cd7f2799d20e5a08487654ae8c99755bad753504)), closes [#2889](https://github.com/cornerstonejs/cornerstone3D/issues/2889) [#2744](https://github.com/cornerstonejs/cornerstone3D/issues/2744) [#2744](https://github.com/cornerstonejs/cornerstone3D/issues/2744) [#2744](https://github.com/cornerstonejs/cornerstone3D/issues/2744) [#2744](https://github.com/cornerstonejs/cornerstone3D/issues/2744) [#2744](https://github.com/cornerstonejs/cornerstone3D/issues/2744) [#2889](https://github.com/cornerstonejs/cornerstone3D/issues/2889)
+
+## [5.9.1](https://github.com/cornerstonejs/cornerstone3D/compare/v5.9.0...v5.9.1) (2026-09-09)
+
+**Note:** Version bump only for package @cornerstonejs/core
+
+# [5.9.0](https://github.com/cornerstonejs/cornerstone3D/compare/v5.8.9...v5.9.0) (2026-09-09)
+
+### Features
+
+- **logging:** replace console calls with hierarchical loglevel loggers named cs3d.<package>.<path>.<file> ([#2865](https://github.com/cornerstonejs/cornerstone3D/issues/2865)) ([49dc03d](https://github.com/cornerstonejs/cornerstone3D/commit/49dc03dcb7badce594525e9285368964a38aa174))
+
+## [5.8.9](https://github.com/cornerstonejs/cornerstone3D/compare/v5.8.8...v5.8.9) (2026-09-03)
+
+**Note:** Version bump only for package @cornerstonejs/core
+
+## [5.8.8](https://github.com/cornerstonejs/cornerstone3D/compare/v5.8.7...v5.8.8) (2026-09-02)
+
+### Bug Fixes
+
+- **core:** keep the camera finite for single-slice volumes ([#2858](https://github.com/cornerstonejs/cornerstone3D/issues/2858)) ([62f78e6](https://github.com/cornerstonejs/cornerstone3D/commit/62f78e63f4e397eaff48528061415b3dfb12300a))
+
+## [5.8.7](https://github.com/cornerstonejs/cornerstone3D/compare/v5.8.6...v5.8.7) (2026-09-01)
+
+**Note:** Version bump only for package @cornerstonejs/core
+
+## [5.8.6](https://github.com/cornerstonejs/cornerstone3D/compare/v5.8.5...v5.8.6) (2026-09-01)
+
+### Bug Fixes
+
+- **StackViewport:** cancel the pending scroll timer when the viewport is disabled ([#2884](https://github.com/cornerstonejs/cornerstone3D/issues/2884)) ([bff87eb](https://github.com/cornerstonejs/cornerstone3D/commit/bff87eb02c1d79372eca75f09ff186aca2c62e76))
+
+## [5.8.5](https://github.com/cornerstonejs/cornerstone3D/compare/v5.8.4...v5.8.5) (2026-09-01)
+
+**Note:** Version bump only for package @cornerstonejs/core
+
+## [5.8.4](https://github.com/cornerstonejs/cornerstone3D/compare/v5.8.3...v5.8.4) (2026-08-27)
+
+**Note:** Version bump only for package @cornerstonejs/core
+
+## [5.8.3](https://github.com/cornerstonejs/cornerstone3D/compare/v5.8.2...v5.8.3) (2026-08-27)
+
+**Note:** Version bump only for package @cornerstonejs/core
+
+## [5.8.2](https://github.com/cornerstonejs/cornerstone3D/compare/v5.8.1...v5.8.2) (2026-08-17)
+
+**Note:** Version bump only for package @cornerstonejs/core
+
+## [5.8.1](https://github.com/cornerstonejs/cornerstone3D/compare/v5.8.0...v5.8.1) (2026-08-14)
+
+**Note:** Version bump only for package @cornerstonejs/core
+
+# [5.8.0](https://github.com/cornerstonejs/cornerstone3D/compare/v5.7.3...v5.8.0) (2026-08-13)
+
+**Note:** Version bump only for package @cornerstonejs/core
+
+## [5.7.3](https://github.com/cornerstonejs/cornerstone3D/compare/v5.7.2...v5.7.3) (2026-08-11)
+
+**Note:** Version bump only for package @cornerstonejs/core
+
+## [5.7.2](https://github.com/cornerstonejs/cornerstone3D/compare/v5.7.1...v5.7.2) (2026-08-11)
+
+### Bug Fixes
+
+- **core:** return CSS pixels from the CPU worldToCanvas/canvasToWorld ([bdd1dd3](https://github.com/cornerstonejs/cornerstone3D/commit/bdd1dd3c26c50802c1c5ac3530eec30ecdbbc39d)), closes [#2849](https://github.com/cornerstonejs/cornerstone3D/issues/2849)
+
+## [5.7.1](https://github.com/cornerstonejs/cornerstone3D/compare/v5.7.0...v5.7.1) (2026-08-11)
+
+**Note:** Version bump only for package @cornerstonejs/core
+
+# [5.7.0](https://github.com/cornerstonejs/cornerstone3D/compare/v5.6.14...v5.7.0) (2026-08-07)
+
+### Features
+
+- **tools:** touch interaction support with dispatcher fallback to mouse callbacks ([#2815](https://github.com/cornerstonejs/cornerstone3D/issues/2815)) ([667ba22](https://github.com/cornerstonejs/cornerstone3D/commit/667ba224254bfcbc4d14b87a2cdc9a9d5dae0807))
+
+## [5.6.14](https://github.com/cornerstonejs/cornerstone3D/compare/v5.6.13...v5.6.14) (2026-08-06)
+
+**Note:** Version bump only for package @cornerstonejs/core
+
 ## [5.6.13](https://github.com/cornerstonejs/cornerstone3D/compare/v5.6.12...v5.6.13) (2026-08-03)
 
 **Note:** Version bump only for package @cornerstonejs/core